added coins and food order assignment

+++ coins.java
@@ -3,12 +3,12 @@ public class coins {
     public static void main(String[] args) {
         Scanner read = new Scanner(System.in);
         System.out.println("Enter amount in cents: ");
-        double cents = read.nextInt();
+        int cents = read.nextInt();
 
-        double quarters = cents/25;
-        double dimes = cents%25/10;
-        double nickels = dimes%25%10/5;
-        double nickelsLeft = cents%5;
+        int quarters = cents/25;
+        int dimes = cents%25/10;
+        int nickels = dimes%25%10/5;
+        int nickelsLeft = cents%5;
 
         System.out.println("Quarters: " + quarters);
         System.out.println("Dimes: " + dimes);
